Added MySQL datasource, issue 29
