+++ src/metrics/mysql_metric.ts
@@ -1,0 +1,5 @@
+import {SqlMetric} from "./sql_metric";
+
+export class MysqlMetric extends SqlMetric {
+
+}
